apply styling suggestions
