+++ Source/Calling/WireCallCenterV3+Events.swift
@@ -326,18 +326,6 @@ extension WireCallCenterV3 {
     }
 }
 
-extension WireCallCenterV3 {
-
-    func callDidDegrade(conversationId: UUID, degradedUser: ZMUser) {
-        closeCall(conversationId: conversationId, reason: .securityDegraded)
-
-        if let previousSnapshot = callSnapshots[conversationId] {
-            callSnapshots[conversationId] = previousSnapshot.updateDegradedUser(degradedUser)
-        }
-    }
-    
-}
-
 private extension Set where Element == ZMUser {
 
     var avsClients: Set<AVSClient> {
